+++ src/main/webapp/META-INF/maven/com.vr/Controller/pom.properties
@@ -1,5 +1,5 @@
 #Generated by Maven Integration for Eclipse
-#Tue Jan 23 12:31:39 KST 2024
+#Wed Jan 24 14:31:28 KST 2024
 version=1.0.0-BUILD-SNAPSHOT
 groupId=com.vr
 m2e.projectName=v31
