--- src/main/webapp/resources/css/sujuletter.css
@@ -1,119 +1,0 @@
-@charset "UTF-8";
-
-
-#print_page{
-	background: white;
-}
-
-#print {
-    position: relative;
-   top: -648px;
-    left: 630px;
-    width: 70px;
-    height: 25px;
-    border-radius: 5px;
-}
-
-#close{
-    position: relative;
-    top: -648px;
-    left: 630px;
-    width: 70px;
-    height: 25px;
-    border-radius: 5px;
-}
-
-#send {
-    position: relative;
-   top: -672px;
-    left: 553px;
-    width: 70px;
-    height: 25px;
-    border-radius: 5px;
-}
-#ment{
-    position: relative;
-    top: 30px;
-    margin-left: 17px;
-}
-
-#aa{
-	position: relative;
-	top: 5px;
-}
-.susutext_title{
-	text-align: center;
-}
-
-html{
-    position: relative;
-    top: 30px;
-}
-
-#total_money{
-	position: relative;
-    left: 160%;
-    font-size: 20px;
-    border-bottom: 1px solid black;
-    width: 131px;
-}
-
-body{
-	margin: auto;
-	width: 21cm;
-    height: 14.8cm;
-    border:1px solid black;
-}
-
-.sujuletter {
-    border-bottom: 1px solid black;
-    width: 40%;
-    margin-bottom: 10px;
-    padding-bottom: 0px;
-    margin-left: 20px;
-    margin-top: 10px;
-    
-}
-#baljuletter {
-    display: inline-block;
-    position: relative;
-    left: 390px;
-    top: -160px;
-    
-}
-
-.subtitle {
-    width: 70px;
-    height: 25px;
-text-align-last: justify;
-}
-
-.input{
-	width: 110px;
-	font-size: 15px;
-}
-
-#company, #company_boss, #registration_number{
-	text-align: center;
-}
-#money{
-	    position: relative;
-    left: -370px;
-    top: 10px;
-}
-#suju_ment1 {
-    position: relative;
-    font-size: 20px;
-}
-#suju_ment2{
-	position: relative;
-    top: 30px;
-    left: -223px;
-    font-size: 20px;
-}
-#suju_table{
-    position: relative;
-    width: 756px;
-    top: -130px;
-    margin-left: 19px;
-}
