+++ src/main/webapp/resources/css/sujuletter1.css
@@ -1,8 +1,98 @@
 @charset "UTF-8";
 
+
 textarea {
 	width: 100%;
     height: 150px;
     border: none;
     resize: none;
+}
+
+
+#print_page{
+	background: white;
+	position: relative;
+    top: 370px;
+}
+
+
+#aa{
+	position: relative;
+	top: 5px;
+}
+.susutext_title{
+	text-align: center;
+}
+
+html{
+    position: relative;
+    top: 30px;
+}
+
+#total_money {
+    position: relative;
+    left: 142%;
+    font-size: 20px;
+    border-bottom: 1px solid black;
+    width: 200px;
+    letter-spacing: 3px;
+}
+
+body{
+	margin: auto;
+	width: 21cm;
+    height: 14.8cm;
+}
+
+.sujuletter {
+    border-bottom: 1px solid black;
+    width: 40%;
+    margin-bottom: 10px;
+    padding-bottom: 0px;
+    margin-left: 20px;
+    margin-top: 10px;
+    
+}
+#baljuletter {
+    display: inline-block;
+    position: relative;
+    left: 390px;
+    top: -160px;
+    
+}
+
+.subtitle {
+    width: 70px;
+    height: 25px;
+text-align-last: justify;
+}
+
+.input{
+	width: 110px;
+	font-size: 15px;
+}
+
+#company, #company_boss, #registration_number{
+	text-align: center;
+}
+#money{
+	    position: relative;
+    left: -370px;
+    top: 10px;
+}
+#suju_ment1 {
+    position: relative;
+    font-size: 20px;
+}
+#suju_ment2{
+	position: relative;
+    top: 30px;
+    left: -223px;
+    font-size: 20px;
+}
+#suju_table{
+    position: relative;
+    width: 756px;
+    top: -130px;
+    margin-left: 19px;
 }
